+++ Libraries/Appboy_iOS_SDK.framework/Headers/ABKLocationManager.h
@@ -1,10 +1,11 @@
 #import <Foundation/Foundation.h>
+#import <CoreLocation/CoreLocation.h>
 
 @class ABKServerConfig;
 
 NS_ASSUME_NONNULL_BEGIN
 
-@interface ABKLocationManager : NSObject
+@interface ABKLocationManager : NSObject <CLLocationManagerDelegate>
 
 /*!
  * Use ABKEnableAutomaticLocationCollectionKey to enable automatic location tracking.
@@ -17,6 +18,12 @@ NS_ASSUME_NONNULL_BEGIN
  * For more information, please refer to Appboy.h.
  */
 @property (readonly) BOOL enableGeofences;
+
+/*!
+ * Use ABKDisableAutomaticGeofenceRequestsKey to disable automatic geofence requests.
+ * For more information, please refer to requestGeofencesWithLongitude:latitude: in Appboy.h
+ */
+@property (readonly) BOOL disableAutomaticGeofenceRequests;
 
 /*!
  * Calling this method will log a location using the regular location provider if a location is reported in under